mesa-gl: Use of mesa-gl assumes you have libmali-xlnx
If a program wants OpenGL, but not GLE you still need files from libmali
in order to compile.  Add necessary links to header files in the -dev package.

Since we now depend on libmali-xlnx, we also now have a machine specific
package.

+++ meta-xilinx-core/recipes-graphics/mesa/mesa-gl_%.bbappend
@@ -1,5 +1,15 @@
 PACKAGECONFIG:append = " dri3 gallium"
 
-do_install:append:zynqmp () {
-    rm -rf ${D}${includedir}/KHR/*
+# If we're using libmali-xlnx, then we need to bring it in for the KHR/khrplatform.h file
+DEPENDS .= "${@' libmali-xlnx' if d.getVar('PREFERRED_PROVIDER_virtual/libgles1') == 'libmali-xlnx' else ''}"
+RDEPENDS:libgl-mesa-dev .= "${@' libmali-xlnx-dev' if d.getVar('PREFERRED_PROVIDER_virtual/libgles1') == 'libmali-xlnx' else ''}"
+
+do_install:append () {
+    if ${@'true' if d.getVar('PREFERRED_PROVIDER_virtual/libgles1') == 'libmali-xlnx' else 'false'} ; then
+        rm -rf ${D}${includedir}/KHR/*
+    fi
 }
+
+# If we require libmali-xlnx, this becomes MACHINE_ARCH specific
+DEFAULT_PACKAGE_ARCH := "${PACKAGE_ARCH}"
+PACKAGE_ARCH = "${@'${MACHINE_ARCH}' if d.getVar('PREFERRED_PROVIDER_virtual/libgles1') == 'libmali-xlnx' else '${DEFAULT_PACKAGE_ARCH}'}"
